+++ earth2studio/io/kv.py
@@ -110,7 +110,9 @@ class KVBackend:
 
         adjusted_coords, mapping = convert_multidim_to_singledim(coords)
 
-        self.coords = self.coords | adjusted_coords
+        for c, v in adjusted_coords.items():
+            if c not in self.coords:
+                self.coords[c] = v
 
         # Add any multidim coordinates that were expelled above
         for k in mapping:
@@ -184,18 +186,17 @@ class KVBackend:
 
         for xi, name in zip(x, array_name):
             if name not in self.root:
-                self.add_array(coords, array_name, data=xi.to(self.device))
+                self.add_array(coords, array_name)
 
-            else:
-                # Get indices as list of arrays and set torch tensor
-                self.root[name][
-                    np.ix_(
-                        *[
-                            np.where(np.isin(self.coords[dim], value))[0]
-                            for dim, value in adjusted_coords.items()
-                        ]
-                    )
-                ] = xi.to(self.device)
+            # Get indices as list of arrays and set torch tensor
+            self.root[name][
+                np.ix_(
+                    *[
+                        np.where(np.isin(self.coords[dim], value))[0]
+                        for dim, value in adjusted_coords.items()
+                    ]
+                )
+            ] = xi.to(self.device)
 
     def to_xarray(self, **xr_kwargs: Any) -> xarray.Dataset:
         """