+++ CHANGELOG.md
@@ -11,12 +11,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Added derived TCWV diagnostic model
+
 ### Changed
 
 - Removed AIFS-ENS invariants from model input / outputs and moved into model wrapper
 - CDS and NCAR lexica have been updated with variables for AIFS-ENS
 - Switch grib reads in HRRR data sources from xarray to pygrib to fix memory leak
 - Updated AIFS model to 1.1 checkpoint version
+- Updated GFS lexicon to include composite reflectivity
 
 ### Deprecated
 
@@ -25,6 +28,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Made tp in HRRR and GFS units m for package consistency
+- Fixed batched inference support for AIFS and AIFS ENS
 
 ### Security